requirements.txt
numpy
pandas
tqdm
librosa
Fire
python-dotenv
pyyaml
h5py

psutil
ipython
flask
better_exchook

pytest
